Consolidate config related functionality into the config package
