--- config/advanced.go
@@ -1,7 +1,0 @@
-package config
-
-// Advanced settings
-type Advanced struct {
-	IncludeCountry    bool `json:"include_country"`
-	UseDomainFronting bool `json:"use_domain_fronting"`
-}
