--- config/notifications.go
@@ -1,8 +1,0 @@
-package config
-
-// Notifications settings
-type Notifications struct {
-	Enabled                bool `json:"enabled"`
-	NotifyOnTestCompletion bool `json:"notify_on_test_completion"`
-	NotifyOnNews           bool `json:"notify_on_news"`
-}
